working haversine base -- not working json parser free

+++ part2/GenData/src/generate_coords.h
@@ -3,5 +3,5 @@
 f64 generate_coords(FILE *flex_out,
                     FILE *ans_out,
                     const Method method,
-                    const u32 seed,
-                    const u32 n_pairs);
+                    const u64 seed,
+                    const u64 n_pairs);
